+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## master
 
+- Add `auth: {token: '...'}` option to `createCable` to pass the initial authentiation token. ([@palkan][])
+
 - Add `transportConfigurator` parameter to `createCable` to perform arbitrary transport modifications (like, fetching the initial token) before opening a connection. ([@palkan][])
 
 ## 1.0.0 (2024-12-26)